AbstractTagsInput extracted
